+++ app/scripts/controllers/point-controller.js
@@ -181,29 +181,19 @@ angular.module('lizard-nxt')
 
 
     $scope.getIconClass = function (str) {
-
-      console.log('[F] $scope.sayHello(str) WHERE str =', str);
-
       switch (str) {
-
       case 'overflow':
         return 'icon-overflow';
-
       case 'pumpstation_sewerage':
         return 'icon-pumpstation-diesel';
-
       case 'bridge':
         return 'icon-bridge';
-
       case 'bridge-draw':
         return 'icon-bridge';
-
       case 'bridge-fixed':
         return 'icon-bridge';
-
       default:
         return 'icon-' + str;
-
       }
     };
   }