+++ app/scripts/services/layer-group-service.js
@@ -114,6 +114,9 @@ angular.module('lizard-nxt')
         }
       });
 
+      // Sort them
+      this._layers = sortLayers(layers);
+
     }
 
     LayerGroup.prototype = {
@@ -128,7 +131,7 @@ angular.module('lizard-nxt')
       */
       toggle: function (map) {
         if (!this._initiated) {
-          this._initializeLayers(this._layers, this.temporal);
+          this._initializeLayers(this._layers);
           this._initiated = true;
         }
 
@@ -466,21 +469,176 @@ angular.module('lizard-nxt')
        * @description delegates initialization of leaflet layers to other
        *              functions.
        */
-      _initializeLayers: function () {
-        angular.forEach(this._layers, function (layer) {
+      _initializeLayers: function (layers) {
+        angular.forEach(layers, function (layer) {
           layer.initializeLayer();
         });
       },
 
       _toggleLayers: function (map, layers, active) {
-        var method = active ? 'add' : 'remove';
-        angular.forEach(layers, function (layer) {
-          layer[method](map);
-        });
+        if (active && layers.length > 0) {
+          addLayersRecursively(map, layers, 0);
+        }
+        else {
+          angular.forEach(layers, function (layer) {
+            layer.remove(map);
+          });
+        }
         if (this._opacity) {
           this.setOpacity(this._opacity);
         }
       }
+    };
+
+    /**
+     * @function
+     * @memberof app.LayerGroup
+     * @param  {array} Array of nxt layers
+     * @return {array} Array of object sorted by property loadOrder in
+     *                 descending order.
+     * @description Sorts layers by descending loadOrder
+     */
+    var sortLayers = function (layers) {
+      layers.sort(function (a, b) {
+        if (a.loadOrder > b.loadOrder) {
+          return -1;
+        }
+        if (a.loadOrder < b.loadOrder
+          || a.loadOrder === null) {
+          return 1;
+        }
+        // a must be equal to b
+        return 0;
+      });
+      return layers;
+    };
+
+    /**
+     * @function
+     * @memberof app.LayerGroup
+     * @param  {object} map Leaflet map to add layers to
+     * @param  {array} Array of nxt layers
+     * @param  {int} i index to start from
+     * @description Adds the layers with the loadorder of layers[i]. Catches
+     *              the returned promises and calls itself with the nxt index.
+     *              When all layers are loaded it adds a listener to the last
+     *              layer with the highest loadOrder.
+     */
+    var addLayersRecursively = function (map, layers, i) {
+      var currentLoadOrder = layers[i].loadOrder;
+      // Wrap contains the promises and the nxt index.
+      var wrap = loadLayersByLoadOrder(map, layers, i, currentLoadOrder);
+      // If there is more, wait for these layers to resolve
+      // and start over with the remaining layers.
+      if (wrap.i < layers.length) {
+        startOverWhenDone(wrap.promises, map, layers, wrap.i);
+      }
+      // When done, add listener to the last layer with the max loadOrder
+      // that is drawn on the map.
+      else if (layers.length > 1) {
+        var index = getIndexOfLeadingLayer(layers);
+        if (typeof(index) === 'number') {
+          addLoadListenersToLayer(map, layers, index);
+        }
+      }
+    };
+
+
+    /**
+     * @function
+     * @memberof app.LayerGroup
+     * @param  {object} map Leaflet map to add layers to.
+     * @param  {array} layers Array of nxt layers.
+     * @param  {int} i index to start from.
+     * @param  {inte} loadOrder Current load order to add layers.
+     * @return {object} next index and list of promises that resolve when layer
+     *                       is fully loaded.
+     * @description Adds the layers from index i with the given loadorder to the
+     *              map. Returns the current index and a list of promises for
+     *              all the added layers when a layer with a lower loadorder is
+     *              found.
+     */
+    var loadLayersByLoadOrder = function (map, layers, i, loadOrder) {
+      // Add all layers with the current load order
+      var promises = [];
+      while (i < layers.length
+        && layers[i].loadOrder === loadOrder) {
+        promises.push(layers[i].add(map));
+        i++;
+      }
+      return {
+        i: i,
+        promises: promises
+      };
+    };
+
+    /**
+     * @function
+     * @memberof app.LayerGroup
+     * @param  {array} layers Array of nxt layers.
+     * @return {int} Index of the last layer with the highest loadOrder.
+     * @description Loops through the sorted layers and returns the index of the
+     *              last layer in the array with the highest loadOrder.
+     */
+    var getIndexOfLeadingLayer = function (layers) {
+      var index;
+      var highestLoadingOrder = 0;
+      for (var i = 0; i < layers.length; i++) {
+        if (layers[i].tiled
+          && (layers[i].loadOrder > highestLoadingOrder
+          || layers[i].loadOrder === highestLoadingOrder)) {
+          index = i;
+          highestLoadingOrder = index;
+        }
+      }
+      return index;
+    };
+
+    /**
+     * @function
+     * @memberof app.LayerGroup
+     * @param  {array} Array of promises.
+     * @param  {object} map Leaflet map to add layers to.
+     * @param  {array} layers Array of nxt layers.
+     * @param  {int} i index to start from.
+     * @description Takes a list of promises and calls addLayersRecursively when
+     *              all promises have resolved.
+     */
+    var startOverWhenDone = function (promises, map, layers, i) {
+      $q.all(promises).then(function () {
+        addLayersRecursively(map, layers, i);
+      });
+    };
+
+    /**
+     * @function
+     * @memberof app.LayerGroup
+     * @param  {object} map Leaflet map to add layers to.
+     * @param  {array} layers Array of nxt layers.
+     * @param  {int} i index to start from.
+     * @description Adds listeners that call when load starts and finished to
+     *              the layer at index i of layers. Callbacks remove layers of
+     *              the map after index i when load starts and adds layers after
+     *              index i recursively when load finishes.
+     */
+    var addLoadListenersToLayer = function (map, layers, i) {
+      var layer = layers[i];
+      var j = i + 1;
+
+      var removeAllAfterI = function () {
+        for (j; j < layers.length; j++) {
+          layers[j].remove(map);
+        }
+      };
+
+      var reAdd = function () {
+        addLayersRecursively(map, layers, i + 1);
+      };
+
+      layer._leafletLayer.off('load');
+      layer._leafletLayer.off('loading');
+      layer._leafletLayer.on('loading', removeAllAfterI);
+      layer._leafletLayer.on('load', reAdd);
     };
 
     return LayerGroup;